Fix: パス探索コマンドが正常に動作
About: #71
